pyproject.toml
[tool.poetry]
name = "eiden-feedback"
version = "1.1.0"
description = ""
authors = ["HowardLovekraft <117375241+HowardLovekraft@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.group.dev]
optional = true

[tool.poetry.dependencies]
python = "^3.10"
aiogram = {extras = ["i18n"], version = "^3.4.1"}
python-dotenv = "^1.0.1"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
